requirements.txt
streamlit
PyPDF2
langchain
openai
tiktoken
faiss-cpu
pillow
pikepdf
imgbeddings
numpy
langchain-anthropic
google-api-python-client
streamlit-pages